requirements.txt
django>=5.0
djangorestframework
psycopg2-binary
python-dotenv
pytest
pytest-django